requirements.txt
Django==4.1.7
djangorestframework==3.14.0
PyYAML==6.0
requests==2.28.2
django-cors-headers==3.14.0
rsa==4.9
pycryptodome==3.18.0